move Build struct to sev-types
