+++ sev/tests/api.rs
@@ -2,7 +2,9 @@
 
 use sev_types::platform::Version;
 
-use sev::{certs::sev::Usage, firmware::Firmware, Build};
+use sev::{certs::sev::Usage, firmware::Firmware};
+
+use sev_types::Build;
 
 #[cfg_attr(not(all(has_sev, feature = "dangerous_tests")), ignore)]
 #[test]
